third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-basic.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
description("Test if requestMIDIAccess is available.");

shouldBeDefined("navigator.requestMIDIAccess");

var promise = navigator.requestMIDIAccess();
shouldBeDefined("promise");

shouldBeDefined("promise.then");

promise.then(function(){}, function(){});

</script>
</body>
</html>
