third_party/WebKit/LayoutTests/webmidi/requestmidiaccess-basic-expected.txt

Test if requestMIDIAccess is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.requestMIDIAccess is defined.
PASS promise is defined.
PASS promise.then is defined.
PASS successfullyParsed is true

TEST COMPLETE
